Changes:  - fixed issue #24, crash when creating a new project with as project store the    directory of an old already existing project
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@1883 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ src/squadt/project_manager.cpp
@@ -46,7 +46,16 @@ namespace squadt {
         write();
       }
       else {
-        read();
+        try {
+          read();
+        }
+        catch (...) {
+          /* Project description file is probably broken */
+          import_directory(l);
+
+          /* Create initial project description file */
+          write();
+        }
       }
     }
     else {
@@ -72,7 +81,9 @@ namespace squadt {
         import_directory(*i);
       }
       else {
-        import_file(*i);
+        if ((*i).leaf() != settings_manager::project_definition_base_name) {
+          import_file(*i);
+        }
       }
 
       ++i;
@@ -339,7 +350,7 @@ namespace squadt {
     path           destination_path = store / path(d.empty() ? s.leaf() : d);
     processor::ptr p                = processor::create(*this);
 
-    if (s != destination_path) {
+    if (s != destination_path && exists(destination_path)) {
       copy_file(s, destination_path);
     }
